remove `src/map/layers/geo-mixin.js`
